# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel C-sending",
        "description": "L5 Swagger OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000"
        }
    ],
    "paths": {
        "/api/campaign": {
            "post": {
                "tags": [
                    "Campaign"
                ],
                "description": "save sending data",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "config",
                                    "type",
                                    "status",
                                    "paramerters",
                                    "destination",
                                    "template"
                                ],
                                "properties": {
                                    "config": {
                                        "type": "object"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "campaign_uuid": {
                                        "type": "integer"
                                    },
                                    "template": {
                                        "type": "string"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "receivers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "integer"
                                                },
                                                "parameters": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "username": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "destination": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "campaign": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "_id": {
                                                                    "type": "string"
                                                                },
                                                                "template": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                },
                                                                "type": {
                                                                    "type": "string"
                                                                },
                                                                "config": {
                                                                    "type": "object"
                                                                },
                                                                "created_at": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "receiver": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "_id": {
                                                                    "type": "integer"
                                                                },
                                                                "campaign_uuid": {
                                                                    "type": "integer"
                                                                },
                                                                "destination": {
                                                                    "type": "integer"
                                                                },
                                                                "status": {
                                                                    "type": "integer"
                                                                },
                                                                "parameters": {
                                                                    "type": "integer"
                                                                },
                                                                "created_at": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/update-status": {
            "post": {
                "tags": [
                    "receiver"
                ],
                "description": "update status",
                "operationId": "update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multiPart/form-data": {
                            "schema": {
                                "required": [
                                    "receiver_id",
                                    "status"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/processed-receivers": {
            "get": {
                "tags": [
                    "receiver"
                ],
                "description": "get processed receiver",
                "operationId": "receiver",
                "parameters": [
                    {
                        "name": "get_all",
                        "in": "path"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "processed_receiver": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "_id": {
                                                                    "type": "string"
                                                                },
                                                                "campaign_uuid": {
                                                                    "type": "integer"
                                                                },
                                                                "destination": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                },
                                                                "parameters": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "contact_first_name": {
                                                                                "type": "string"
                                                                            },
                                                                            "contact_middle_name": {
                                                                                "type": "string"
                                                                            },
                                                                            "contact_last_name": {
                                                                                "type": "string"
                                                                            },
                                                                            "contact_phone": {
                                                                                "type": "integer"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                },
                                                                "created_at": {
                                                                    "type": "string"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}